Perbaikan permission

+++ app/Http/Controllers/BarangController.php
@@ -10,6 +10,8 @@ class BarangController extends Controller
     // 1. Tampilkan semua data barang
     public function index()
     {
+        $this->cekAkses('barang_index'); // 🔒 PASANG GEMBOK Akses Lihat
+
         $barangs = Barang::orderBy('kode_barang', 'asc')->paginate(20);
         return view('barang.index', compact('barangs'));
     }
@@ -17,12 +19,18 @@ class BarangController extends Controller
     // 2. Form tambah barang
     public function create()
     {
+        abort_if(auth()->user()->role !== 'admin', 403, 'FORBIDDEN 🛑 : Hanya Admin yang bisa menyimpan barang!');
+        $this->cekAkses('barang_index'); // 🔒 PASANG GEMBOK Akses Tambah
+
         return view('barang.create');
     }
 
     // 3. Proses simpan data barang baru
     public function store(Request $request)
     {
+        
+        $this->cekAkses('barang_index'); // 🔒 PASANG GEMBOK Akses Simpan
+
         $request->validate([
             'kode_barang' => 'required|unique:barangs',
             'nama_barang' => 'required',
@@ -46,6 +54,9 @@ class BarangController extends Controller
     // 4. Form edit barang
     public function edit($id)
     {
+        abort_if(auth()->user()->role !== 'admin', 403, 'FORBIDDEN 🛑 : Hanya Admin yang bisa mengedit barang!');
+        $this->cekAkses('barang_index'); // 🔒 PASANG GEMBOK Akses Edit
+
         $barang = Barang::findOrFail($id);
         return view('barang.edit', compact('barang'));
     }
@@ -53,6 +64,9 @@ class BarangController extends Controller
     // 5. Proses update data barang
     public function update(Request $request, $id)
     {
+        abort_if(auth()->user()->role !== 'admin', 403, 'FORBIDDEN 🛑 : Hanya Admin yang bisa mengubah barang!');
+        $this->cekAkses('barang_index'); // 🔒 PASANG GEMBOK Akses Update
+
         $barang = Barang::findOrFail($id);
 
         $request->validate([
@@ -76,6 +90,9 @@ class BarangController extends Controller
     // 6. Hapus barang
     public function destroy($id)
     {
+        abort_if(auth()->user()->role !== 'admin', 403, 'FORBIDDEN 🛑 : Hanya Admin yang bisa menghapus barang!');
+        $this->cekAkses('barang_index'); // 🔒 PASANG GEMBOK Akses Hapus
+
         $barang = Barang::findOrFail($id);
         $barang->delete();
